requirements.txt
fastapi[standard]
uvicorn
httpx
python-dotenv
pydantic
